splash update
