+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:portfolio/constants/app_constants.dart';
 import 'package:portfolio/providers/theme_notifier.dart';
 import 'package:portfolio/themes/app_theme.dart';
-import 'package:portfolio/views/portfolio.dart';
+import 'package:portfolio/views/splash/splash_view.dart';
 
 void main() {
   runApp(const ProviderScope(child: MyApp()));
@@ -22,7 +22,7 @@ class MyApp extends ConsumerWidget {
         title: AppConstants.appName,
         debugShowCheckedModeBanner: false,
         theme: AppTheme.getAppTheme(context, ref.watch(themeProvider)),
-        home: const PortFolio(),
+        home: const SplashView(),
       ),
     );
   }
