+++ lib/views/screens/contacts_view/contact_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:portfolio/responsive/responsive_layout.dart';
 import 'package:portfolio/views/screens/contacts_view/contact_desktop_view.dart';
-import 'package:portfolio/views/screens/contacts_view/widgets/contact_mobile_view.dart';
+import 'package:portfolio/views/screens/contacts_view/contact_mobile_view.dart';
 
 class ContactView extends StatelessWidget {
   const ContactView({super.key});
