+++ lib/views/screens/project_details_view/project_details_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ProjectDetailsScreen extends StatelessWidget {
+  const ProjectDetailsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
