+++ lib/views/screens/project_details_view/widgets/project_details_widget.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ProjectDetailsWidget extends StatelessWidget {
+  const ProjectDetailsWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
