+++ lib/views/screens/project_details_view/widgets/project_image_widget.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ProjectDetailWidget extends StatelessWidget {
+  const ProjectDetailWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
