+++ lib/views/screens/work_view/work_desktop_view.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:portfolio/styles/app_styles.dart';
 import 'package:portfolio/themes/app_colors.dart';
+import 'package:portfolio/views/screens/work_view/widgets/work_view_widgets.dart';
 
 class WorkDesktopView extends StatelessWidget {
   const WorkDesktopView({super.key});
@@ -12,7 +13,13 @@ class WorkDesktopView extends StatelessWidget {
       mainAxisSize: MainAxisSize.min,
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        Text("Explore My Latest works", style: AppStyles.headlineMedium),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            WorkViewWidgets.exploreMyLatestWorks(),
+            WorkViewWidgets.seeProjectsButton()
+          ],
+        ),
         SizedBox(
           height: 0.03.sh,
         ),
